% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_columns_class.R
\name{add_columns}
\alias{add_columns}
\title{Add columns}
\usage{
add_columns(new_columns, by, ...)
}
\arguments{
\item{new_columns}{(data.frame) A data.frame (rhs) to be left-joined with the annotation table (lhs).}

\item{by}{(character) A character vector of variables to join by.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{add_columns} object with the following \code{output} slots:
\tabular{ll}{
\code{updated} \tab          (annotation_table) The updated annotations as an annotation_table object. \cr
}
}
\description{
A wrapper around dplyr::left_join. Adds columns to an annotation table by performing a left-join with an input data.frame. (annotations on the left).
}
