% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_class.R
\name{annotation_table}
\alias{annotation_table}
\title{An annotation source}
\usage{
annotation_table(annotations, tag, id_column, ...)
}
\arguments{
\item{annotations}{(data.frame) Data frame of imported annotations.}

\item{tag}{(character) A (short) character string that is used to represent this table e.g. in column names, or source columns.}

\item{id_column}{(name)}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{annotation_table} object with the following \code{output} slots:
\tabular{ll}{
\code{annotations} \tab          (data.frame) Data frame of imported annotations. \cr
}
}
\description{
Base class for annotation sources.
}
