% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_source_class.R
\name{cd_source}
\alias{cd_source}
\title{Import annotation source}
\usage{
cd_source(
  input_file,
  compounds_file,
  tag = "CD",
  cd_version = 3.1,
  add_cols = list(),
  ...
)
}
\arguments{
\item{input_file}{(character) The file annotations are imported from.}

\item{compounds_file}{(character) The file of compounds output by Compound Discoverer.}

\item{tag}{(character) An abbreviation used to identify this source in annotation tables. The default is \code{"CD"}.}

\item{cd_version}{(numeric) The version of Compound Discoverer used to generate the input files. The default is \code{3.1}.}

\item{add_cols}{(list) A (named) list of additional columns to be added to the table and populated with the provided value. The default is \code{list()}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{cd_source} object with the following \code{output} slots:
\tabular{ll}{
\code{imported} \tab          (lcms_table) The imported annotations as an lcms_table object. \cr
}
}
\description{
Imports an annotation source.
}
