% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sources.R
\name{combine_tables}
\alias{combine_tables}
\title{Combine annotation sources (tables)}
\usage{
combine_tables(
  source_list,
  matching_columns = NULL,
  keep_cols = NULL,
  tag_ids = FALSE,
  source_col = "annotation_table",
  exclude_cols = NULL,
  ...
)
}
\arguments{
\item{source_list}{(list) A list of annotation sources to be combined.}

\item{matching_columns}{(list, NULL) A named list of columns to be created by merging columns from individual sources. E.g. list("cake" = c("cake_1", "cake2"))will create a new column called "cake" by merging the "cake_1" and "cake_2" if they are found in the input tables. The default is \code{NULL}.}

\item{keep_cols}{(character, NULL) A list of column names to keep in the combined table (padded with NA) if detected in one of the input tables. The default is \code{NULL}.}

\item{tag_ids}{(logical) Create a column of new identifiers for each annotation by merging the id from each source with the annotation tag. e.g.an annotation with id = "A1" for source tagged "XY" is given a new id "A1_XY" to prevent duplicate IDs from multiple sources. The default is \code{FALSE}.}

\item{source_col}{(character) The column name that will be created to contain  a tag to indicate which source the annotation originated from. The default is \code{"annotation_table"}.}

\item{exclude_cols}{(NULL, character) Colum names to be excluded from the merged annotation table. The default is \code{NULL}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{combine_tables} object with the following \code{output} slots:
\tabular{ll}{
\code{combined_table} \tab          (annotation_table) The annotation tabel after combining the input tables. \cr
}
}
\description{
Annotation tables are joined and matching columns merged.
}
