% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_records_class.R
\name{.collapse}
\alias{.collapse}
\title{Collapse
for use with combine_records
returns a function to collapse a vector into a string using the provided separator}
\usage{
.collapse(separator)
}
\description{
Collapse
for use with combine_records
returns a function to collapse a vector into a string using the provided separator
}
