% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_records_class.R
\name{.count}
\alias{.count}
\title{number of annotations}
\usage{
.count()
}
\description{
for use with combine records
Adds a new column indicating the number of annotations that match the given
grouping variable.
}
