% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_records_class.R
\name{.median}
\alias{.median}
\title{Median
for use with collapse_records
returns a function for calculating the median value, excluding NA}
\usage{
.median()
}
\description{
Median
for use with collapse_records
returns a function for calculating the median value, excluding NA
}
