% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_records_class.R
\name{.mode}
\alias{.mode}
\title{Mode
returns the most common value, excluding NA.
first value when sorted (min for numeric)}
\usage{
.mode()
}
\description{
Mode
returns the most common value, excluding NA.
first value when sorted (min for numeric)
}
