% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_records_class.R
\name{.nothing}
\alias{.nothing}
\title{do nothing to the annotations}
\usage{
.nothing()
}
\description{
for use with combine records
A pass-through function to allow some annotation table columns to remain
unchanged.
}
