% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_records_class.R
\name{.select_max}
\alias{.select_max}
\title{select annotations with a maximum}
\usage{
.select_max(max_col, use_abs = FALSE)
}
\description{
for use with combine_records
returns a function to select the value of a list based on the index of the maximum in a second list
}
