% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_records_class.R
\name{.unique}
\alias{.unique}
\title{unique values
for use with combine_records
reduces a list to unique values, then collapses using the separator}
\usage{
.unique(separator, na_string = "NA", digits = 6)
}
\description{
unique values
for use with combine_records
reduces a list to unique values, then collapses using the separator
}
