% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_levels_class.R
\name{filter_levels}
\alias{filter_levels}
\title{Filter by factor levels}
\usage{
filter_levels(column_name, levels, mode = "exclude", ...)
}
\arguments{
\item{column_name}{(character) The column name to filter.}

\item{levels}{(character) The levels to filter by.}

\item{mode}{(character) Filter mode. Allowed values are limited to the following: \itemize{\item{\code{"exclude"}: The specified levels are removed from the annotation table.}\item{\code{"include"}: Only the specified levels are retained in the annotation table.}} The default is \code{"exclude"}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{filter_levels} object with the following \code{output} slots:
\tabular{ll}{
\code{filtered} \tab          (annotation_source) Annotation_source after filtering. \cr
\code{flags} \tab          (data.frame) A list of flags indicating which annotations were removed. \cr
}
}
\description{
Removes (or includes) annotations such that the named column excludes (or includes) the specified levels.
}
