% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_count_class.R
\name{id_counts}
\alias{id_counts}
\title{id counts}
\usage{
id_counts(id_column, count_column = "id_counts", ...)
}
\arguments{
\item{id_column}{(character) Column name of the variable ids in variable_meta.}

\item{count_column}{(character) The name of the new column to store the counts in. The default is \code{"id_counts"}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{id_counts} object with the following \code{output} slots:
\tabular{ll}{
\code{updated} \tab          (annotation_table) The input annotation source with the newly generated column. \cr
}
}
\description{
Adds the number of times an identical identifier is present to each record.
}
