% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_lookup_class.R
\name{id_db_lookup}
\alias{id_db_lookup}
\title{ID lookup by database}
\usage{
id_db_lookup(
  annotation_column,
  database_column,
  id_database,
  include = NULL,
  tag = NULL,
  not_found = NA,
  ...
)
}
\arguments{
\item{annotation_column}{(character) The annotation table column name to use as the reference for searching the database e.g. "HMBD_ID".}

\item{database_column}{(character) The database column to search for matches to the values in annoation_column.}

\item{id_database}{(data.frame) A database of annotations and identifiers to be searched.}

\item{include}{(character, NULL) The name of the database columns to be added to the annotations. If NULL, all columns are retained. The default is \code{NULL}.}

\item{tag}{(character, NULL) A string appended to the column names from the database. Used to distinguish columns from different databases with identical column names.If tag = NULL then the column names are not changed. The default is \code{NULL}.}

\item{not_found}{(character, numeric, logical, NULL) The returned value when there are no matches. The default is \code{NA}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{id_db_lookup} object with the following \code{output} slots:
\tabular{ll}{
\code{updated} \tab          (annotation_table) The input annotation sources is updated with matching columns from the is database. \cr
}
}
\description{
Search a database (data.frame) for annotation matches based on values in a specified column.
}
\details{
This object makes use of functionality from the following packages:\itemize{\item{\code{dplyr}}}
}
\references{
Wickham H, François R, Henry L, Müller K (2022).
\emph{dplyr: A Grammar of Data Manipulation}.
R package version 1.0.10, \url{https://CRAN.R-project.org/package=dplyr}.
}
