% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_annotation_class.R
\name{import_annotation}
\alias{import_annotation}
\title{Import annotation}
\usage{
import_annotation(...)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{import_annotation} object with the following \code{output} slots:
\tabular{ll}{
\code{imported} \tab          (annotation_source) The imported annotations as an annotation_source object. \cr
}
}
\description{
Imports an annotation source as part of a model.sequence.
}
