% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTox_RT_fcns.R
\name{import_MTox_RT}
\alias{import_MTox_RT}
\title{Import annotation}
\usage{
import_MTox_RT(
  add_cols = list(),
  tag = "MTox_RT",
  source = c("rt_only", "spectral_lib"),
  assay,
  ...
)
}
\arguments{
\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{import_annotation} object with the following \code{output} slots:
\tabular{ll}{
\code{imported} \tab          (annotation_source) The imported annotations as an annotation_source object. \cr
}
}
\description{
Imports an annotation source as part of a model.sequence.
}
