% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_annotation_class.R
\name{import_source,annotation_source-method}
\alias{import_source,annotation_source-method}
\title{Import annotation source}
\usage{
\S4method{import_source}{annotation_source}(M)
}
\description{
An `annotation_source` is imported and parsed into an `annotation_table` object.
}
