% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_lookup_class.R
\name{kegg_lookup}
\alias{kegg_lookup}
\title{Search Kegg for identifiers}
\usage{
kegg_lookup(
  id_column,
  kegg_db = "compound",
  external_db = "pubchem",
  out_column = "cid",
  mode = "kegg_to_external",
  ...
)
}
\arguments{
\item{id_column}{(character) The name of the column containing identifiers to search the database for.}

\item{kegg_db}{(character) kegg database. Allowed values are limited to the following: \itemize{\item{\code{"compound"}: Small molecule database.}\item{\code{"glycan"}: Glycan database.}\item{\code{"drug"}: Drug database.}} The default is \code{"compound"}.}

\item{external_db}{(character) external database. Allowed values are limited to the following: \itemize{\item{\code{"pubchem"}: PubChem CID.}\item{\code{"chebi"}: ChEBI ID.}} The default is \code{"pubchem"}.}

\item{out_column}{(character) The name of the newly generated column. The default is \code{"cid"}.}

\item{mode}{(character) Lookup mode. Allowed values are limited to the following: \itemize{\item{\code{"kegg_to_external"}: Search the kegg database for matching kegg identifiers and return external identifiers.}\item{\code{"external_to_kegg"}: Search the kegg database for matching external identifiers and return kegg identifiers.}} The default is \code{"kegg_to_external"}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{kegg_lookup} object with the following \code{output} slots:
\tabular{ll}{
\code{updated} \tab          (annotation_table) Annotation_table a new column of compound identifiers. \cr
}
}
\description{
Searches the Kegg database to obtain external identifiers e.g. PubChem CID and vice-versa
}
