% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_class.R
\name{lcms_source}
\alias{lcms_source}
\title{An annotation source}
\usage{
lcms_source(input_file, ...)
}
\arguments{
\item{input_file}{(character) The file annotations are imported from.}

\item{...}{Additional slots and values passed to \code{struct_class}.}

\item{tag}{(character) A (short) character string that is appended to columns names when merging this annotation source with other sources. Used to identify columns that are from this source in the merged output.}

\item{add_cols}{(list) A list of additional columns to be added to the table and populated with the provided value. The default is \code{list()}.}
}
\value{
A  \code{annotation_source} object with the following \code{output} slots:
\tabular{ll}{
\code{annotations} \tab          (data.frame) Data frame of imported annotations. \cr
}
}
\description{
Base class for annotation sources.
}
