% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidmaps_lookup_class.R
\name{lipidmaps_lookup}
\alias{lipidmaps_lookup}
\title{LipidMaps api lookup}
\usage{
lipidmaps_lookup(
  search_column,
  output_columns,
  return_columns,
  context,
  context_item,
  ...
)
}
\arguments{
\item{search_column}{(character) The column of the annotation table that contains search terms.}

\item{output_columns}{(character) The column names to store the results. Must be the same length as return_columns.}

\item{return_columns}{(character) Return columns. Allowed values are limited to the following: \itemize{\item{\code{""}: The name of the columns to return from the results of the search. Must be value(s) from the following list: .}\item{\code{"compound"}: Input,regno,lm_id,name,sys_name,synonyms,abbrev,abbrev_chains,core,main_class,sub_class,exactmass,formula,inchi,inchi_key,hmdb_id,chebi_id,pubchem_cid,smiles,kegg_id.}\item{\code{"gene"}: Gene_id,lmp_id,gene_name,gene_symbol,gene_synonyms,chromosome,map_location,summary,taxid,species,species_long.}\item{\code{"protein"}: Uniprot_id,lmp_id,gene_id,gene_name,gene_symbol,taxid,species,species_long,mrna_id,refseq_id,protein_gi,protein_entry,protein_name,seq_length,seq.}}}

\item{context}{(character) The search API context. Must be one of compound,gene,protein.}

\item{context_item}{(character) Context item. Allowed values are limited to the following: \itemize{\item{\code{""}: The context item being searched. Must be:.}\item{\code{"compound"}: Lm_id,formula,inchi_key,pubchem_cid,hmdb_id,kegg_id,chebi_id,smiles,abbrev,abbrev_chains.}\item{\code{"gene"}: Lmp_id,gene_id,gene_name,gene_symbol,tax_id.}\item{\code{"protein"}: Lmp_id,gene_id,gene_name,gene_symbol,tax_id,mrna_id,refseq_id,protein_gi,protein_entry,protein_name,uniprot_id.}}}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{lipidmaps_lookup} object with the following \code{output} slots:
\tabular{ll}{
\code{updated} \tab          (annotation_source) The annotation_source after being updated with the results of the search. \cr
\code{results_table} \tab          (data.frame) The returned search results in full. \cr
}
}
\description{
Search the LipidMaps database using the API
}
