% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_source_class.R
\name{ls_source}
\alias{ls_source}
\title{Import annotation source}
\usage{
ls_source(input_file, add_cols = list(), ...)
}
\arguments{
\item{input_file}{(character) The file annotations are imported from.}

\item{add_cols}{(list) A (named) list of additional columns to be added to the table and populated with the provided value. The default is \code{list()}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{ls_source} object with the following \code{output} slots:
\tabular{ll}{
\code{imported} \tab          (annotation_table) The imported annotations as an annotation_table object. \cr
}
}
\description{
Imports an annotation source.
}
