% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_class.R
\name{model_apply,model_seq,annotation_table-method}
\alias{model_apply,model_seq,annotation_table-method}
\title{Apply model sequence}
\usage{
\S4method{model_apply}{model_seq,annotation_table}(M, D)
}
\description{
Applies a model sequence to an annotations table
}
