% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem_lookup_class.R
\name{pubchem_lookup}
\alias{pubchem_lookup}
\title{Compound ID lookup via pubchem}
\usage{
pubchem_lookup(
  annotation_column,
  out_column = "cid",
  search_by = "name",
  return_value = "cids",
  property = "InChIKey",
  ...
)
}
\arguments{
\item{annotation_column}{(character) The column name to use as the reference for searching the database e.g. "HMBD_ID". If length > 1, then priority is given to columns left-to-right.}

\item{out_column}{(character) The name of the column to store the returned value for the annotations. The default is \code{"cid"}.}

\item{search_by}{(character) The PubChem domain to search for matches to the annotation_column. The default is \code{"name"}.}

\item{return_value}{(character) The values to return if a match is found. The default is \code{"cids"}.}

\item{property}{(character) Property value to return if a match is found, and return_value = "property". The default is \code{"InChIKey"}.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{pubchem_lookup} object with the following \code{output} slots:
\tabular{ll}{
\code{updated} \tab          (annotation_source) The updated annotation source. \cr
}
}
\description{
Uses the PubChem API to search for CID based on the input annotation column.
}
