% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicates}
\usage{
remove_duplicates(mode, columns, ...)
}
\arguments{
\item{mode}{(character) Mode. Allowed values are limited to the following: \itemize{\item{\code{"include"}: The named columns are considered when searching for duplicates.}\item{\code{"exclude"}: The named columns are not considered when searching for duplicates.}}}

\item{columns}{(character) The column names considered (or not) when searching for duplicates.}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{remove_duplicates} object with the following \code{output} slots:
\tabular{ll}{
\code{updated} \tab          (annotation_table) The updated annotations as an annotation_table object. \cr
}
}
\description{
A wrapper around dplyr::distinct to remove duplicate rows from the annotation table.
}
