% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_records_class.R
\name{.mean}
\alias{.mean}
\title{Mean
for use with collapse_records
returns a function for calculating the mean value, excluding NA}
\usage{
.mean()
}
\description{
Mean
for use with collapse_records
returns a function for calculating the mean value, excluding NA
}
