% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_records_class.R
\name{.mode}
\alias{.mode}
\title{report modal value for an annotation}
\usage{
.mode(ties = FALSE, ignore_na = TRUE)
}
\description{
Returns the most common value, excluding NA.
If ties == TRUE then all tied values are returned, else the first value in
a sorted unique list is return (equal to min if numeric)
If ignore_na == FALSE then NA are included when searching for the modal value
and placed last if ties = FALSE (values are returned preferentially over NA)
}
