% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_records_class.R
\name{.prioritise}
\alias{.prioritise}
\title{prioritise
#' for use with combine records
reduces a list by prioritizing values according to the input
if multiple matches to the priority then collapse using separator}
\usage{
.prioritise(match_col, priority, separator, no_match = NA, na_string = "NA")
}
\description{
prioritise
#' for use with combine records
reduces a list by prioritizing values according to the input
if multiple matches to the priority then collapse using separator
}
