% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTox_RT_class.R
\name{import_MTox_RT}
\alias{import_MTox_RT}
\title{Import MTox RT library}
\usage{
import_MTox_RT(
  add_cols = list(),
  tag = "MTox_RT",
  source = c("rt_only", "spectral_lib"),
  assay,
  ...
)
}
\arguments{
\item{add_cols}{(list) A named list of values to add to all records. The default is \code{list()}.}

\item{tag}{(character) The tag name used to indentify this annotation source in e.g. column names. The default is \code{"MTox_RT"}.}

\item{source}{(character) The source names to retain in the imported library. Values can be: "spectral_lib", "rt_only" or c("spectral_lib","rt_only"). The default is \code{c("rt_only", "spectral_lib")}.}

\item{assay}{(character) The assay names to retain in the imported library (HILIC_NEG, HILIC_POS, LIPIDS_POS, LIPIDS_NEG).}

\item{...}{Additional slots and values passed to \code{struct_class}.}
}
\value{
A  \code{import_MTox_RT} object with the following \code{output} slots:
\tabular{ll}{
\code{imported} \tab          (lcms_table) The imported annotations as an lcms_table object. \cr
}
}
\description{
Imports the MTox retention time library as an annotation source.
}
